requirements.txt
cx-oracle
mysql-connector
schedule